requirements.txt
numpy==1.16.4
matplotlib==3.3.4
scikit-image==0.17.2
opencv-python==3.4.2.16
pillow==8.3.1


